+++ src/container/main/components/Footer/index.style.tsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+import { AppBar } from '@mui/material';
+
+export const FooterStyled = styled(AppBar)`
+    background-color: black;
+    position: relative;
+    bottom: 0;
+`;
