+++ src/container/main/components/Header/index.style.tsx
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+import { AppBar } from '@mui/material';
+
+export const HeaderStyled = styled(AppBar)`
+    background-color: black;
+`;
