+++ src/container/main/components/Header/index.tsx
@@ -1,23 +1,23 @@
 import React from 'react';
-import { getNavigationsValue } from '@brojs/cli';
+import { getNavigationValue } from '@brojs/cli';
 import { Box, Typography, Toolbar } from '@mui/material';
-import RefButton from '../../../components/ref-button';
-import IcButton from '../../../components/icon-button';
+import RefButton from '../../../../components/ref-button';
+import IcButton from '../../../../components/icon-button';
 import { HeaderStyled } from './index.style';
 
 const navigations: Array<{ name: string; href: string }> = [
     {
         name: 'Главная',
-        href: getNavigationsValue('sber_web.main')
+        href: getNavigationValue('sber_web.main')
     },
     {
         name: 'Детальная информация',
-        href: getNavigationsValue('sber_web.detail')
+        href: getNavigationValue('sber_web.detail')
     },
 
     {
         name: 'Редактировать',
-        href: getNavigationsValue('sber_web.edit')
+        href: getNavigationValue('sber_web.edit')
     }
 ];
 
@@ -25,7 +25,7 @@ const Header = (): React.ReactElement => {
     return (
         <HeaderStyled position="static">
             <Toolbar>
-                <IcButton href={getNavigationsValue('sber_web.main')} />
+                <IcButton href={getNavigationValue('sber_web.main')} />
                 <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
                     Что? Где? Когда?
                 </Typography>
