+++ src/container/main/components/Main/index.style.tsx
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const MainStyled = styled.main`
+    // Какие-то стили для содержимого страницы (например шрифты)
+    flex-grow: 1;
+`;
