+++ src/container/main/components/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Header from './header';
-import Main from './main';
-import Footer from './footer';
+import Header from './Header';
+import Main from './Main';
+import Footer from './Footer';
 import { LayoutStyled } from './index.style';
 
 const Layout = (): React.ReactElement => {
